Optimize cache contains check
Profiling revealed that we were spending a lot of time in
`bitset.equals`. This is because our cache was indexed by
`bitset.popcnt`, so when we wanted to see if a particular (bitset,
state) pair was in the cache, we ended up doing a linear search over a
large array, doing lots of `bitset.equals` calls in the process.

This patch introduces `bitset.hash`, which produces a much better
quality hash than `bitset.popcnt`. Switching the cache to use this hash
function greatly improves performance (by over 150x on many of the
slowest tests).