Add "jump to first error" button in visualization
This patch adds a "jump to first error" button to the visualization,
which scrolls to the first visible bad linearization point, so that
users do not have to manually scroll through the visualization looking
for red lines. Clicking on the button jumps to the first visible error,
so it respects the currently selected partition, if any. If no partition
is selected, clicking on the button has the effect of selecting the
partition containing the error point, to make it easier to see what is
going on.

This patch also tweaks the legend so it is floating.