Fix bug with handling of zero-duration operations
The code that handled conversion from `[]Operation` to `[]entry` sorted
operations by time, but it didn't properly handle the case where there
were multiple operations with zero duration (call timestamp equal to
return timestamp) and the same timestamp, which can happen if the clock
is not fine-grained enough. For a subset of entries where the timestamp
is the same, the call entries need to be ordered before the return
entries.