Improve handling of large timestamp values
Timestamps in Porcupine are represented as int64s. However, in the
visualization, we use the JavaScript Number type (float64). Previously,
we were converting from one to the other through JSON, and interpreting
the JSON as raw JavaScript, resulting in a lossy conversion from int64
to float64 (the same thing would have happened with `JSON.parse()`).
This works fine for small numbers, but when using, for example,
nanosecond-level timestamps, these values can be larger than
JavaScript's `Number.MAX_SAFE_INTEGER`, causing problems downstream in
the visualization code.

This patch applies a simple approach to solving this problem. The
visualization doesn't need the original timestamp values (except for
tooltip display purposes) because it warps time anyways to produce a
compact display, so we apply a monotonic map to the timestamp values to
"compress" them into a smaller space, ensuring that they're a lot
smaller than `Number.MAX_SAFE_INTEGER`.

Thanks to Yun-Sheng Chang <yschang@mit.edu> for the bug report.